Move back to trait as non-abstract functions
