+++ std/shared/src/main/scala/cats/effect/std/syntax/AllSyntax.scala
@@ -17,8 +17,4 @@
 package cats.effect
 package std.syntax
 
-trait AllSyntax
-    extends BackpressureSyntax
-    with SupervisorSyntax
-    with QueueSyntax
-    with DequeueSyntax
+trait AllSyntax extends BackpressureSyntax with SupervisorSyntax
